add some refactors
