+++ src/constants/texts.ts
@@ -1,4 +1,5 @@
 export const TEXTS = {
+  // UserCard (src/components/sections/header/UserCard.tsx)
   changePassword: "Thay đổi mật khẩu",
   resetPassword: "Đặt lại mật khẩu",
   logout: "Đăng xuất",
@@ -9,12 +10,14 @@ export const TEXTS = {
   login: "Đăng nhập",
   register: "Tạo tài khoản",
 
+  // LoginView (src/modules/auth/login/views/LoginView.tsx)
   termsPrefix: "Bằng cách đăng nhập, bạn đồng ý với",
   termsOfUse: "Điều khoản sử dụng",
   and: "và",
   privacyPolicy: "Chính sách bảo mật",
   our: "của chúng tôi.",
 
+  // RatingSection (src/components/product/RatingSection.tsx)
   reviewAndRating: "Nhận xét và đánh giá",
   noRatings: "Chưa có đánh giá",
   ratings: "đánh giá",
@@ -24,8 +27,10 @@ export const TEXTS = {
   toReview: "để nhận xét và đánh giá sản phẩm này",
   writeReview: "Viết đánh giá",
 
+  // Login page (src/app/(auth)/login/page.tsx)
   loginTitle: "Thạch Âm - Đăng nhập",
 
+  // RegisterGridView (src/components/auth/RegisterGridView.tsx)
   emailExists: "Email này đã được đăng ký",
   useAnotherEmail: "Vui lòng sử dụng email khác hoặc đăng nhập",
   registerSuccess: "Đã đăng ký tài khoản thành công",
@@ -39,6 +44,7 @@ export const TEXTS = {
   alreadyHaveAccount: "Đã có tài khoản?",
   loginNow: "Đăng nhập ngay",
 
+  // RecoveryGridView (src/components/auth/RecoveryGridView.tsx)
   checkInbox: "Kiểm tra hộp thư của bạn",
   ifEmailValid: "Nếu email hợp lệ, bạn sẽ nhận được liên kết khôi phục.",
   checkSpam: "Đừng quên kiểm tra cả hộp thư rác (spam).",
@@ -53,8 +59,10 @@ export const TEXTS = {
   requestFailed: "Gửi yêu cầu thất bại",
   errorOccurred: "Đã có lỗi xảy ra. Vui lòng thử lại sau.",
 
+  // Password recovery page (src/app/(auth)/password-recovery/page.tsx)
   passwordRecoveryTitle: "Thạch Âm - Đặt lại mật khẩu",
 
+  // LoginGridView (src/components/auth/LoginGridView.tsx)
   welcomeBack: "Chào mừng bạn trở lại",
   loginToContinue: "Đăng nhập để tiếp tục hành trình của bạn",
   orLoginWithEmail: "Hoặc đăng nhập bằng email",
@@ -63,4 +71,7 @@ export const TEXTS = {
   registerNow: "Đăng ký ngay",
   invalidEmailOrPassword: "Email hoặc mật khẩu không chính xác",
   checkLoginInfo: "Vui lòng kiểm tra lại thông tin đăng nhập",
+
+  // GeneratedAvatar (src/components/sections/header/GeneratedAvatar.tsx)
+  avatarError: "Không thể lấy dữ liệu avatar",
 };
